Refactor leave repo and add leave balance/services
Unify and extend leave management: replaced specialized leave/leave-balance interfaces with a single ILeaveRepository, removed legacy ILeaveBalanceRepository and related classes, and moved leave-balance CRUD into LeaveRepository. Rename and fix typos for LeaveTableCOnfig -> LeaveTableConfig (and related property/method names), correct casing for email/password/redirect_link, and adjust method names (getLeaveTableConfig). Add reminder and incrementing features: new repository methods (getAllRequestNeedsReminder, incrementAllEmployeeLeaveRequest, create/update/delete/get leave balance) and register IncrementLeaveBalanceService and LeaveRequestReminderService as hosted services. Update handlers/controllers/commands/DTOs accordingly (use ILeaveRepository, adjust CreateLeaveRequestCommand usage, UpdateLeaveRequestDto defaults and types, map update logic and attachment handling, fix email templates and date/day amount casting). Remove several obsolete files and make minor model/DbContext adjustments to match renamed types.

+++ HRManagement.Api/Application/Commands/LeaveManagementCommands/Helper/LeaveTemplate.cs
@@ -19,7 +19,7 @@ namespace HRManagement.Api.Application.Commands.LeaveManagementCommands.Helper
 
             Leave ID: {leaveRequest.LeaveId}
             Category: {MappingHelper.leaveTypeFromInt((int)leaveRequest!.LeaveType).ToString()}
-            Date: {leaveRequest.LeaveStartDate} - {leaveRequest.LeaveStartDate.AddDays(leaveRequest.DayAmount)}
+            Date: {leaveRequest.LeaveStartDate} - {leaveRequest.LeaveStartDate.AddDays((double)leaveRequest.DayAmount)}
 
             Please review and approve the request using the link below.
             {link}
@@ -66,7 +66,7 @@ namespace HRManagement.Api.Application.Commands.LeaveManagementCommands.Helper
 
             Leave ID: {leaveRequest.LeaveId}
             Category: {MappingHelper.leaveTypeFromInt((int)leaveRequest!.LeaveType).ToString()}
-            Date: {leaveRequest.LeaveStartDate} - {leaveRequest.LeaveStartDate.AddDays(leaveRequest.DayAmount)}
+            Date: {leaveRequest.LeaveStartDate} - {leaveRequest.LeaveStartDate.AddDays((double)leaveRequest.DayAmount)}
 
             Please review and approve the request using the link below:
             {link}
